Add PDF quote preview functionality and discounts modalIDEC-42-previsualizar-cotizacion-en-pdf (#17)
* feat: Agregar modal para añadir descuentos en NuevaCotizacion

* chore: Agregar @react-pdf/renderer y sus dependencias a  package.json y package-lock.json

* feat: Agregar componente QuotePDF con estilos y fuentes personalizadas

* feat: Agregar página de previsualización de cotización en PDF

* refactor: QuoteState y agregar lógica de cálculo de subtotal, impuestos y total

* Merge branch 'main' into IDEC-42-previsualizar-cotizacion-en-pdf

* refactor: eliminar la interfaz Quote y el tipo QuoteStatus no utilizados

* feat: mejorar la previsualización de cotizaciones con carga y spinner, y ajustar el manejo de datos

* feat: agregar campo customer_name en el estado de cotización y en la interfaz Quote

* feat: actualizar formato de fecha y moneda para incluir nuevos estilos y opciones personalizadas

* feat: agregar funcionalidad para previsualizar cotizaciones en una nueva pestaña

+++ src/utils/currency.ts
@@ -5,7 +5,7 @@ type FormatStyle = 'short' | 'long'
 
 export function formatCurrency(
   amount: number,
-  style: FormatStyle = 'short',
+  style: FormatStyle = 'long',
   locale = DEFAULT_LOCALE,
   currency = DEFAULT_CURRENCY,
   customOptions?: Intl.NumberFormatOptions
